delete decodeDirectly api for simplicity
