+++ modules/barcode/include/opencv2/barcode.hpp
@@ -86,16 +86,6 @@ public:
     CV_WRAP bool detectAndDecode(InputArray img, CV_OUT std::vector<std::string> &decoded_info, CV_OUT
                                  std::vector<BarcodeType> &decoded_type, OutputArray points = noArray()) const;
 
-    /** @brief Decode without detects
-     *
-     * @param img grayscale or color (BGR) image containing barcode.
-     * @param decoded_info UTF8-encoded output of string or empty string if the codes do not contain barcode.
-     * @param decoded_type vector of BarcodeType, specifies the type of these barcodes
-    */
-    CV_WRAP bool
-    decodeDirectly(InputArray img, CV_OUT String &decoded_info, CV_OUT BarcodeType &decoded_type) const;
-
-
 //protected:
 //    struct Impl;
 //    Ptr <Impl> p;
