+++ modules/barcode/src/barcode.cpp
@@ -6,7 +6,6 @@
 
 #include "precomp.hpp"
 #include <opencv2/barcode.hpp>
-#include "decoder/bardecode.hpp"
 #include "decoder/ean13_decoder.hpp"
 #include "detector/bardetect.hpp"
 
@@ -50,6 +49,74 @@ static void updatePointsResult(OutputArray points_, const vector<Point2f> &point
             points_.release();
         }
     }
+}
+
+class BarDecode
+{
+public:
+    void init(const Mat &src, const vector<Point2f> &points);
+
+    const vector<Result> &getDecodeInformation()
+    { return result_info; }
+
+    bool decodeMultiplyProcess();
+
+private:
+    vector<vector<Point2f>> src_points;
+    Mat original;
+    vector<Result> result_info;
+};
+
+void BarDecode::init(const cv::Mat &src, const std::vector <cv::Point2f> &points) {
+    //CV_TRACE_FUNCTION();
+    original = src.clone();
+    CV_Assert(!points.empty());
+    CV_Assert((points.size() % 4) == 0);
+    src_points.clear();
+    for (size_t i = 0; i < points.size(); i += 4) {
+        vector <Point2f> tempMat{points.cbegin() + i, points.cbegin() + i + 4};
+        if (contourArea(tempMat) > 0.0) {
+            src_points.push_back(tempMat);
+        }
+    }
+    CV_Assert(!src_points.empty());
+}
+
+bool BarDecode::decodeMultiplyProcess() {
+    class ParallelBarCodeDecodeProcess : public ParallelLoopBody {
+    public:
+        ParallelBarCodeDecodeProcess(Mat &inarr_, vector <vector<Point2f>> &src_points_,
+                                     vector <Result> &decoded_info_)
+                : inarr(inarr_), decoded_info(decoded_info_), src_points(src_points_) {
+            for (size_t i = 0; i < src_points.size(); ++i) {
+                decoder.push_back(std::unique_ptr<AbsDecoder>(new Ean13Decoder()));
+            }
+        }
+
+        void operator()(const Range &range) const
+
+        CV_OVERRIDE
+        {
+            CV_Assert(inarr.channels() == 1);
+            Mat gray = inarr.clone();
+            for (int i = range.start; i < range.end; i++) {
+                Mat bar_img;
+                cutImage(gray, bar_img, src_points[i]);
+                decoded_info[i] = decoder[i]->decodeImg(bar_img, src_points[i]);
+            }
+        }
+
+    private:
+        Mat &inarr;
+        vector <Result> &decoded_info;
+        vector <vector<Point2f>> &src_points;
+        vector <std::unique_ptr<AbsDecoder>> decoder;
+    };
+    result_info.clear();
+    result_info.resize(src_points.size());
+    ParallelBarCodeDecodeProcess parallelDecodeProcess{original, src_points, result_info};
+    parallel_for_(Range(0, int(src_points.size())), parallelDecodeProcess);
+    return !result_info.empty();
 }
 
 //struct BarcodeDetector::Impl
@@ -141,23 +208,5 @@ bool BarcodeDetector::detectAndDecode(InputArray img, vector<std::string> &decod
     return ok;
 }
 
-bool BarcodeDetector::decodeDirectly(InputArray img, String &decoded_info, BarcodeType &decoded_type) const
-{
-    Mat inarr;
-    if (!checkBarInputImage(img, inarr))
-    {
-        return false;
-    }
-    Result _decoded_info;
-    std::unique_ptr<AbsDecoder> decoder{new Ean13Decoder()};
-    _decoded_info = decoder->decodeImg(inarr);
-    decoded_info = _decoded_info.result;
-    decoded_type = _decoded_info.format;
-    if (decoded_type == BarcodeType::NONE || decoded_info.empty())
-    {
-        return false;
-    }
-    return true;
-}
 }// namespace barcode
 } // namespace cv
