+++ modules/barcode/src/precomp.hpp
@@ -7,5 +7,6 @@
 
 #include <opencv2/imgproc.hpp>
 #include <numeric>
+#include <string>
 
 #endif
